+++ bluemira/neutronics/make_geometry.py
@@ -3,8 +3,12 @@ import copy
 
 import numpy as np
 import openmc
+from numpy import pi
 
+import bluemira.neutronics.result_presentation as present
 import bluemira.neutronics.volume_functions as vf
+from bluemira.base.constants import raw_uc
+from bluemira.neutronics.params import TokamakGeometry
 
 cells = {}
 surfaces = {}
@@ -12,11 +16,9 @@ surfaces = {}
 # Setting the thickness of divertor below the first wall
 div_clearance = 49.0
 
-# ------------------------------------------------------------------------------------
-
 
 def check_geometry(tokamak_geometry):
-    # Some basic geometry checks
+    """Some basic geometry checks"""
 
     if tokamak_geometry.elong < 1.0:
         raise ValueError("Elongation must be at least 1.0")
@@ -38,23 +40,17 @@ def check_geometry(tokamak_geometry):
         )
 
 
-# ------------------------------------------------------------------------------------
-
-
 def normalizeVec(bisX, bisY):
-    # Normalises a vector
+    """Normalises a vector"""
 
     length = (bisX**2 + bisY**2) ** 0.5
 
     return bisX / length, bisY / length
 
 
-# ------------------------------------------------------------------------------------
-
-
 def get_cone_eqn_from_two_points(p1, p2):
-    # Gets the equation of the OpenMC cone surface from two points
-    # Assumes x0 = 0 and y0 = 0
+    """Gets the equation of the OpenMC cone surface from two points
+    Assumes x0 = 0 and y0 = 0"""
 
     # print("p1, p2:", p1, p2)
 
@@ -85,12 +81,9 @@ def get_cone_eqn_from_two_points(p1, p2):
     return z0, r02
 
 
-# ------------------------------------------------------------------------------------
-
-
 def makeOffsetPoly(oldX, oldY, offset, outer_ccw):
-    # Makes a larger polygon with the same angle from a specified offset
-    # Not mathematically robust for all polygons
+    """Makes a larger polygon with the same angle from a specified offset
+    Not mathematically robust for all polygons"""
 
     num_points = len(oldX)
 
@@ -125,11 +118,8 @@ def makeOffsetPoly(oldX, oldY, offset, outer_ccw):
     return (newX, newY)
 
 
-# ------------------------------------------------------------------------------------
-
-
 def offset_points(points, offset_cm):
-    # Calls makeOffsetPoly with points in the format it expects to get the points of an offset polygon
+    """Calls makeOffsetPoly with points in the format it expects to get the points of an offset polygon"""
 
     old_rs = []
     old_zs = []
@@ -147,33 +137,24 @@ def offset_points(points, offset_cm):
     return np.array(layer_points)
 
 
-# ------------------------------------------------------------------------------------
-
-
 def shift_points(points, shift_cm):
-    # Moves all radii of points outwards by shift_cm
+    """Moves all radii of points outwards by shift_cm"""
 
     points[:, 0] += shift_cm
 
     return points
 
 
-# ------------------------------------------------------------------------------------
-
-
 def elongate(points, adjust_elong):
-    # Adjusts the elongation of the points
+    """Adjusts the elongation of the points"""
 
     points[:, 2] *= adjust_elong
 
     return points
 
 
-# ------------------------------------------------------------------------------------
-
-
-def stretch_r(points, tokamak_geometry, stretch_r_val):
-    # Moves the points in the r dimension away from the major radius by extra_r_cm
+def stretch_r(points, tokamak_geometry: TokamakGeometry, stretch_r_val):
+    """Moves the points in the r dimension away from the major radius by extra_r_cm"""
 
     tokamak_geometry.major_r
 
@@ -184,31 +165,22 @@ def stretch_r(points, tokamak_geometry, stretch_r_val):
     return points
 
 
-# ------------------------------------------------------------------------------------
-
-
 def get_min_r_of_points(points):
-    # Adjusts the elongation of the points
+    """Adjusts the elongation of the points"""
 
     min_r = np.amin(points, axis=0)[0]
 
     return min_r
 
 
-# ------------------------------------------------------------------------------------
-
-
 def get_min_max_z_r_of_points(points):
-    # Adjusts the elongation of the points
+    """Adjusts the elongation of the points"""
 
     min_z = np.amin(points, axis=0)[2]
     max_z = np.amax(points, axis=0)[2]
     max_r = np.amax(points, axis=0)[0]
 
     return min_z, max_z, max_r
-
-
-# ------------------------------------------------------------------------------------
 
 
 def create_inboard_layer(
@@ -219,7 +191,7 @@ def create_inboard_layer(
     layer_name,
     material_lib,
 ):
-    # Creates a layer of inboard cells for scoring
+    """Creates a layer of inboard cells for scoring"""
 
     cells[prefix_for_layer + "_cells"] = []
     surfaces[prefix_for_layer + "_surfs"] = {}
@@ -310,9 +282,6 @@ def create_inboard_layer(
     return
 
 
-# ------------------------------------------------------------------------------------
-
-
 def create_outboard_layer(
     prefix_for_layer,
     prefix_for_layer_behind,
@@ -321,7 +290,7 @@ def create_outboard_layer(
     layer_name,
     material_lib,
 ):
-    # Creates a layer of outboard cells for scoring
+    """Creates a layer of outboard cells for scoring"""
 
     cells[prefix_for_layer + "_cells"] = []
     surfaces[prefix_for_layer + "_surfs"] = {}
@@ -400,9 +369,6 @@ def create_outboard_layer(
     print("Created", layer_name)
 
     return
-
-
-# ------------------------------------------------------------------------------------
 
 
 def create_divertor(div_points, outer_points, inner_points, material_lib):
@@ -636,11 +602,8 @@ def create_divertor(div_points, outer_points, inner_points, material_lib):
     return
 
 
-# ------------------------------------------------------------------------------------
-
-
 def create_plasma_chamber():
-    # Creating the cells that live inside the first wall
+    """Creating the cells that live inside the first wall"""
 
     cells["plasma_inner1"] = openmc.Cell(
         region=-surfaces["meeting_r_cyl"] & +surfaces["divertor_surfs"]["top_surface"],
@@ -677,11 +640,12 @@ def create_plasma_chamber():
     return
 
 
-# ------------------------------------------------------------------------------------
-
-
 def make_geometry(
-    tokamak_geometry, fw_points, div_points, num_inboard_points, material_lib
+    tokamak_geometry: TokamakGeometry,
+    fw_points,
+    div_points,
+    num_inboard_points,
+    material_lib,
 ):
     """
     Create a dictionary of cells
@@ -1093,3 +1057,174 @@ def make_geometry(
     geometry.export_to_xml()
 
     return cells, universe
+
+
+def load_fw_points(tokamak_geometry: TokamakGeometry, save_plots: bool = True):
+    """
+    Load given first wall points,
+        scale them according to the given major and minor radii,
+        then downsample them so that a simplified geometry can be made.
+
+    Currently these are full of magic numbers (e.g. manually chosen integer indices)
+        because it was tailored for one specific model.
+
+    Parameters
+    ----------
+    tokamak_geometry: TokamakGeometry
+
+    Returns
+    -------
+    new_downsampled_fw: points belonging to the first wall
+    new_downsampled_div: points belonging to the divertor
+    num_inboard_points: Number of points in new_downsampled_fw
+                        that belongs to the inboard section.
+
+    Dataflow diagram for this function
+    ----------------------------------
+    blanket_face.npy                divertor_face.npy
+            ↓                           ↓
+    full_blanket_2d_outline         divertor_2d_outline
+            ↓                           ↓
+    (select plasma-facing part)         ↓
+            ↓                           ↓
+    inner_blanket_face(ibf)             ↓
+            ↓                           ↓
+    (downsampling)                  (downsampling)
+            ↓                           ↓
+    downsampled_ibf                 divertor_2d_outline
+            ↓                           ↓
+            →           old_points      ←
+                            ↓
+        (expand points outwards using 'shift_cm')
+                            ↓
+                        new_points
+                            ↓
+                    (elongate, stretch)
+                            ↓
+                        new_points
+                        ↓       ↓
+    new_downsampled_fw  ←       →  new_downsampled_div
+
+    Units
+    -----
+    All units for the diagram above are in cgs
+    """
+    ######## get data ########
+    # Get the geometry from existing .npy files, each is an array of
+    # 3D coordinates of points sampled along the divertor first wall outline.
+    full_blanket_2d_outline = np.load("blanket_face.npy")[0]
+    divertor_2d_outline = np.load("divertor_face.npy")[0]
+    ######## <magic numbers and magic function> ########
+    # that fits only these npy model.
+    # The plasma geometry
+    ex_pts_maj_r = 900.0
+    ex_pts_min_r = 290.0
+    ex_pts_elong = 1.792
+    # Specifying the number of the selected points that define the inboard.
+    num_inboard_points = 6
+    # indices
+    selected_fw_samples = [
+        0,
+        4,
+        8,
+        11,
+        14,
+        17,
+        21,
+        25,
+        28,
+        33,
+        39,
+        42,
+        -1,
+    ]  # sample points
+    selected_div_samples = [
+        72,
+        77,
+        86,
+    ]  # also going to use first and last points from first wall
+    num_points_belongong_to_divertor = len(selected_div_samples)
+
+    def _fix_downsampled_ibf(ds_ibf):
+        """Move the point that is too close to plasma
+        by moving it closer to the central column instead.
+        """
+        ds_ibf[-5][0] = ds_ibf[-5][0] - 25.0
+        return ds_ibf
+
+    ######## </magic numbers and magic function> ########
+    # we will get rid of this whole function later on anyways.
+
+    ######## select the part of the outline facing the plasma ########
+    ibf = inner_blanket_face = full_blanket_2d_outline[52:-2]
+
+    ######## (down)sample existing data ########
+    # blanket
+    downsampled_ibf = raw_uc(ibf[selected_fw_samples], "m", "cm")
+    downsampled_ibf = _fix_downsampled_ibf(downsampled_ibf)
+    # divertor
+    downsampled_divf = raw_uc(divertor_2d_outline[selected_div_samples], "m", "cm")
+
+    ######## Create the full plasma-facing outline by concatenating existing var ########
+    old_points = np.concatenate((downsampled_ibf, downsampled_divf), axis=0)
+
+    print("FW points before adjustment\n", old_points)
+
+    ######## rescale data to fit new geometry. ########
+    # Expand point outwards according to new major radius
+    shift_cm = tokamak_geometry.major_r - ex_pts_maj_r
+    new_points = shift_points(old_points, shift_cm)
+
+    # Adjusting points for elongation and minor radius
+    # This elongation also include an allowance for the minor radius
+    elong_w_minor_r = tokamak_geometry.minor_r / ex_pts_min_r * tokamak_geometry.elong
+    stretch_r_val = tokamak_geometry.minor_r / ex_pts_min_r
+    new_points = elongate(new_points, elong_w_minor_r / ex_pts_elong)
+    new_points = stretch_r(new_points, tokamak_geometry, stretch_r_val)
+
+    ######## split 'new_points' into new_downsampled_* variables ########
+    new_downsampled_fw = new_points[:-num_points_belongong_to_divertor]
+    new_downsampled_div = np.concatenate(
+        (new_points[-(num_points_belongong_to_divertor + 1) :], new_points[:1]), axis=0
+    )
+
+    ######## plotting. ########
+    if save_plots:
+        ######## create parametric variables for plotting smoother lines ########
+        # https://hibp.ecse.rpi.edu/~connor/education/plasma/PlasmaEngineering/Miyamoto.pdf pg. 239 # noqa: W505
+        # R = R0 + a cos(θ + δ sin θ)
+        # where a = minor radius
+        #       δ = triangularity
+        u = tokamak_geometry.major_r  # x-position of the center
+        v = 0.0  # y-position of the center
+        a = tokamak_geometry.minor_r  # radius on the x-axis
+        b = tokamak_geometry.elong * tokamak_geometry.minor_r  # radius on the y-axis
+        tri = tokamak_geometry.triang  # triangularity
+        t = np.linspace(0, 2 * pi, 100)
+
+        with present.PoloidalXSPlot("blanket_face.svg", "Blanket Face") as ax:
+            ax.scatter(full_blanket_2d_outline[:, 0], full_blanket_2d_outline[:, 2])
+
+        with present.PoloidalXSPlot(
+            "all_points_before_after.svg", "Points sampled for making the MCNP model"
+        ) as ax:
+            ax.plot(old_points[:, 0], old_points[:, 2], label="Initial fw points")
+            ax.plot(new_points[:, 0], new_points[:, 2], label="Adjusted fw points")
+            ax.plot(
+                u + a * np.cos(t + tri * np.sin(t)),
+                v + b * np.sin(t),
+                label="Plasma envelope",
+            )  # source envelope
+            ax.legend(loc="upper right")
+
+        with present.PoloidalXSPlot(
+            "selected_pts_inner_blanket_face.svg", "Selected points on the inner blanket"
+        ) as ax:
+            ax.scatter(new_downsampled_fw[:, 0], new_downsampled_fw[:, 2])
+
+        with present.PoloidalXSPlot(
+            "selected_pts_divertor_face.svg", "Selected points on the divertor face"
+        ) as ax:
+            ax.scatter(new_downsampled_div[:, 0], new_downsampled_div[:, 2])
+
+    return new_downsampled_fw, new_downsampled_div, num_inboard_points
